# doc/api.yaml
openapi: 3.0.0
paths:
...
  /album/{id}:
    get:
      operationId: AlbumsController_findOne
      summary: Get single album by id
      description: Get single album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Album not found.
      tags:
        - Album
    put:
      operationId: AlbumsController_update
      summary: Update album information
      description: Update library album information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: The album has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Album not found
      tags:
        - Album
    delete:
      operationId: AlbumsController_delete
      summary: Delete album
      description: Delete album from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The album has been deleted
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Album not found
      tags:
        - Album
  /track:
    get:
      operationId: TracksController_find
      summary: Gets all library tracks list
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '401':
          description: Access token is missing or invalid
      tags:
        - Track
    post:
      operationId: TracksController_create
      summary: Add new track
      description: Add new track information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTrackDto'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. body does not contain required fields
        '401':
          description: Access token is missing or invalid
      tags:
        - Track
  /track/{id}:
    get:
      operationId: TracksController_findOne
      summary: Gets single track by id
      description: Gets single track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. TackId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Track not found.
      tags:
        - Track
    put:
      operationId: TracksController_update
      summary: Update track information
      description: Update library track information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: The track has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Track not found
      tags:
        - Track
    delete:
      operationId: TracksController_delete
      summary: Delete track
      description: Delete track from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The track has been deleted
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Track not found
      tags:
        - Track
...
  /artist/{id}:
    get:
      operationId: ArtistsController_findOne
      summary: Get single artist
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Artist not found.
      tags:
        - Arrtist
    put:
      operationId: ArtistsController_update
      summary: Update artist information
      description: Update library artist information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: The artist has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Artist not found
      tags:
        - Arrtist
    delete:
      operationId: ArtistsController_delete
      summary: Delete artist
      description: Delete artist from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The artist has been deleted
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Artist not found
      tags:
        - Arrtist
  /favs:
    get:
      operationId: FavoritesController_find
      summary: Gets all favorites
      description: Gets all favorites
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '401':
          description: Access token is missing or invalid
      tags:
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      summary: Add track to favorites
      description: Add track to favorite
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseFav'
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '422':
          description: Track not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteTrack
      summary: Delete track from favorites
      description: Delete track from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The track has been deleted
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Favorites
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      summary: Add album to favorites
      description: Add album to favorite
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseFav'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '422':
          description: Album not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteAlbum
      summary: Delete album from favorites
      description: Delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The album has been deleted
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Favorites
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      summary: Add artist to favorites
      description: Add artist to favorite
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseFav'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '422':
          description: Artist not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteArtist
      summary: Delete artist from favorites
      description: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The artist has been deleted
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Favorites
info:
  title: REST Service
  description: >-
    Let's try to create a Home Library Service! Users can create, read, update,
    delete data about Artists, Tracks and Albums, add them to Favorites in their
    own Home Library!
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
  schemas:
...
    Album:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          example: Music for the Jilted Generation
        year:
          type: number
          example: 1994
        artistId:
          type: string
          example: d2992f19-7cac-4a86-9b3d-9d143a42b66d
      required:
        - id
        - name
        - year
        - artistId
...
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          default: null
      required:
        - name
        - year
    Track:
      type: object
      properties:
        id:
          type: string
          example: d9683e62-ce01-4ac4-80c7-23828cd13792
        name:
          type: string
          example: Breathe
        artistId:
          type: string
          example: 97c8d54a-b5e4-4a71-90b6-3786a5bbc094
        albumId:
          type: string
          example: 10c8d54a-v5e4-5a71-90b6-3786a5cec073
        duration:
          type: number
          example: 214
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    AddTrackDto:
      type: object
      properties:
        name:
          type: string
          default: Cosmic Girl
        duration:
          type: number
          default: 241
        artistId:
          default: null
          allOf:
            - $ref: '#/components/schemas/AddTrackDto'
        albumId:
          default: null
          allOf:
            - $ref: '#/components/schemas/AddTrackDto'
      required:
        - name
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          default: Cosmic Girl radio edit
        artistId:
          type: string
          default: null
        albumId:
          type: string
          default: null
        duration:
          type: number
          default: 207
      required:
        - name
        - artistId
        - duration
    Artist:
      type: object
      properties:
        id:
          type: string
          example: 97c8d54a-b5e4-4a71-90b6-3786a5bbc094
        name:
          type: string
          example: Prodigy
        grammy:
          type: boolean
          example: true
      required:
        - id
        - name
        - grammy
...
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    Favorite:
      type: object
      properties:
        artists:
          default:
            - id: 97c8d54a-b5e4-4a71-90b6-3786a5bbc094
              name: Prodigy
              grammy: true
          type: array
          items:
            type: string
        albums:
          example:
            - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              name: Innuendo
              year: 1991
              artistId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          type: array
          items:
            type: string
        tracks:
          example:
            - id: d9683e62-ce01-4ac4-80c7-23828cd13792
              name: breathe
              duration: 214
              artistId: 97c8d54a-b5e4-4a71-90b6-3786a5bbc094
              albumId: null
          type: array
          items:
            type: string
      required:
        - artists
        - albums
        - tracks
    ResponseFav:
      type: object
      properties:
        message:
          type: string
          example:
            message: Track successfully added
      required:
        - message